+++ snowflake-distributed-service-api/src/main/java/com/dwarfeng/sfds/api/integration/subgrade/SnowFlakeLongIdKeyFetcher.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
+@Deprecated
 @Component
 public class SnowFlakeLongIdKeyFetcher implements KeyFetcher<LongIdKey> {