+++ snowflake-distributed-service-api/src/test/java/com/dwarfeng/sfds/api/configuration/GenerateConfiguration.java
@@ -1,0 +1,21 @@
+package com.dwarfeng.sfds.api.configuration;
+
+import com.dwarfeng.sfds.api.integration.subgrade.SnowflakeLongGenerator;
+import com.dwarfeng.sfds.api.integration.subgrade.SnowflakeLongIdKeyGenerator;
+import com.dwarfeng.sfds.stack.service.GenerateService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class GenerateConfiguration {
+
+    @Bean
+    public SnowflakeLongGenerator snowflakeLongGenerator(GenerateService generateService) {
+        return new SnowflakeLongGenerator(generateService);
+    }
+
+    @Bean
+    public SnowflakeLongIdKeyGenerator snowflakeLongIdKeyGenerator(GenerateService generateService) {
+        return new SnowflakeLongIdKeyGenerator(generateService);
+    }
+}