+++ snowflake-distributed-service-impl/src/main/java/com/dwarfeng/sfds/impl/service/GenerateServiceImpl.java
@@ -1,7 +1,7 @@
 package com.dwarfeng.sfds.impl.service;
 
 import com.dwarfeng.sfds.stack.handler.GenerateHandler;
-import com.dwarfeng.sfds.stack.service.LongIdService;
+import com.dwarfeng.sfds.stack.service.GenerateService;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
@@ -11,52 +11,51 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Deprecated
 @Service
-public class LongIdServiceImpl implements LongIdService {
+public class GenerateServiceImpl implements GenerateService {
 
     private final GenerateHandler generateHandler;
 
     private final ServiceExceptionMapper sem;
 
-    public LongIdServiceImpl(GenerateHandler generateHandler, ServiceExceptionMapper sem) {
+    public GenerateServiceImpl(GenerateHandler generateHandler, ServiceExceptionMapper sem) {
         this.generateHandler = generateHandler;
         this.sem = sem;
     }
 
     @Override
-    public long nextLongId() throws ServiceException {
+    public long generateLong() throws ServiceException {
         try {
             return generateHandler.generateLong();
         } catch (Exception e) {
-            throw ServiceExceptionHelper.logParse("生成下一个 Long ID 时发生异常", LogLevel.WARN, e, sem);
+            throw ServiceExceptionHelper.logParse("生成 Long 时发生异常", LogLevel.WARN, e, sem);
         }
     }
 
     @Override
-    public LongIdKey nextLongIdKey() throws ServiceException {
+    public LongIdKey generateLongIdKey() throws ServiceException {
         try {
             return generateHandler.generateLongIdKey();
         } catch (Exception e) {
-            throw ServiceExceptionHelper.logParse("生成下一个 LongIdKey 时发生异常", LogLevel.WARN, e, sem);
+            throw ServiceExceptionHelper.logParse("生成 LongIdKey 时发生异常", LogLevel.WARN, e, sem);
         }
     }
 
     @Override
-    public List<Long> nextLongId(int size) throws ServiceException {
+    public List<Long> generateLong(int size) throws ServiceException {
         try {
             return generateHandler.generateLong(size);
         } catch (Exception e) {
-            throw ServiceExceptionHelper.logParse("生成下一组 Long ID 时发生异常", LogLevel.WARN, e, sem);
+            throw ServiceExceptionHelper.logParse("生成 Long 组时发生异常", LogLevel.WARN, e, sem);
         }
     }
 
     @Override
-    public List<LongIdKey> nextLongIdKey(int number) throws ServiceException {
+    public List<LongIdKey> generateLongIdKey(int size) throws ServiceException {
         try {
-            return generateHandler.generateLongIdKey(number);
+            return generateHandler.generateLongIdKey(size);
         } catch (Exception e) {
-            throw ServiceExceptionHelper.logParse("生成下一组 LongIdKey 时发生异常", LogLevel.WARN, e, sem);
+            throw ServiceExceptionHelper.logParse("生成 LongIdKey 组时发生异常", LogLevel.WARN, e, sem);
         }
     }
 }