+++ snowflake-distributed-service-stack/src/main/java/com/dwarfeng/sfds/stack/service/LongIdService.java
@@ -7,11 +7,19 @@ import com.dwarfeng.subgrade.stack.service.Service;
 import java.util.List;
 
 /**
- * LongID服务。
+ * Long ID 服务。
  *
  * @author DwArFeng
  * @since 0.0.1-alpha
+ * @deprecated 由于命名不合理，该接口已经被 {@link GenerateService} 取代。
  */
+/*
+ * 由于其它项目中仍在在使用该服务，所以需要使用该服务。
+ *   仍在使用该服务的文件：
+ *     - node 模块 spring/application-context-dubbo.xml。
+ */
+@SuppressWarnings("DeprecatedIsStillUsed")
+@Deprecated
 public interface LongIdService extends Service {
 
     /**
@@ -41,12 +49,12 @@ public interface LongIdService extends Service {
     List<Long> nextLongId(int size) throws ServiceException;
 
     /**
-     * 生成下一组 LongIdKey
+     * 生成下一组 LongIdKey。
      *
-     * @param number 生成 ID 的数量。
+     * @param size 生成 ID 的数量。
      * @return 下一组 LongIdKey
      * @throws ServiceException 服务异常。
      * @since 1.4.7
      */
-    List<LongIdKey> nextLongIdKey(int number) throws ServiceException;
+    List<LongIdKey> nextLongIdKey(int size) throws ServiceException;
 }