+++ snowflake-distributed-service-stack/src/main/java/com/dwarfeng/sfds/stack/service/exception/ClockMovedBackwardsException.java
@@ -1,21 +1,30 @@
 package com.dwarfeng.sfds.stack.service.exception;
 
-public class ClockMovedBackwardsException extends Exception {
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
 
-    private static final long serialVersionUID = 2787926000953950500L;
+/**
+ * 时钟回拨异常。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class ClockMovedBackwardsException extends HandlerException {
 
-    public ClockMovedBackwardsException() {
+    private static final long serialVersionUID = 8409003384648965626L;
+
+    private final long duration;
+
+    public ClockMovedBackwardsException(long duration) {
+        this.duration = duration;
     }
 
-    public ClockMovedBackwardsException(String message) {
-        super(message);
-    }
-
-    public ClockMovedBackwardsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ClockMovedBackwardsException(Throwable cause) {
+    public ClockMovedBackwardsException(Throwable cause, long duration) {
         super(cause);
+        this.duration = duration;
+    }
+
+    @Override
+    public String getMessage() {
+        return "Clock moved backwards. Refusing to generate id for " + duration + " milliseconds";
     }
 }